+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D11/DX11blend.cpp
@@ -17,7 +17,6 @@
 
 #include "Direct3D11Headers.h"
 #include "Graphics_Systems/General/GSblend.h"
-#include "Graphics_Systems/General/GScolors.h"
 #include "Graphics_Systems/General/GSprimitives.h"
 
 using namespace enigma::dx11;
