+++ ENIGMAsystem/SHELL/Graphics_Systems/General/GScolors.cpp
@@ -16,6 +16,7 @@
 *** You should have received a copy of the GNU General Public License along
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
+
 #include "GScolors.h"
 #include "GScolor_macros.h"
 
@@ -25,27 +26,13 @@ static inline int min(int x,int y){return x<y ? x:y;}
 static inline int max(int x,int y){return x>y ? x:y;}
 static inline int bclamp(int x){return x > 255 ? 255 : x < 0 ? 0 : x;}
 
-namespace enigma
-{
-	unsigned char currentcolor[4] = {0,0,0,255};
-	int currentblendmode[2] = {0,0};
-	int currentblendtype = 0;
-}
+namespace enigma {
 
-namespace enigma_user
-{
+unsigned char currentcolor[4] = {0,0,0,255};
 
-int draw_get_blend_mode() {
-  return enigma::currentblendmode[0];
-}
+} // namespace enigma
 
-int draw_get_blend_mode_ext(bool src) {
-  return enigma::currentblendmode[(src==true?0:1)];
-}
-
-int draw_get_blend_mode_type() {
-  return enigma::currentblendtype;
-}
+namespace enigma_user {
 
 	int merge_color(int c1,int c2,double amount)
 	{
