+++ ENIGMAsystem/SHELL/Graphics_Systems/None/NONEblend.cpp
@@ -16,8 +16,8 @@
 *** You should have received a copy of the GNU General Public License along
 *** with this code. If not, see <http://www.gnu.org/licenses/>
 **/
+
 #include "Graphics_Systems/General/GSblend.h"
-#include "Graphics_Systems/General/GScolors.h"
 
 namespace enigma_user
 {
